# sfmtp/packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SFMTP API
  version: 1.3.0-phase4
  description: |
    Smart Farm Management & Traceability Platform REST API.

    Conventions (docs/06-api-contracts.md): JSON, snake_case, UTC ISO-8601
    timestamps, UUIDv7 ids, cursor pagination, RFC 9457 problem+json errors.
    Every farm route lives under `/farms/{farm}`; a farm you are not an active
    member of is indistinguishable from one that does not exist (404).

    This file is the contract between backend, web and mobile. The backend
    test `OpenApiCoverageTest` fails if a route is missing here.
servers:
  - url: /api/v1
security:
  - bearer: []
...
paths:
...
  /farms/{farm}:
    parameters:
      - $ref: '#/components/parameters/Farm'
    get:
      tags: [Farms]
      operationId: getFarm
      summary: Get farm
      x-permission: farm.profile.view
      responses:
        '200':
          description: Farm profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Farm' }
        '404': { $ref: '#/components/responses/Problem' }
    put:
      tags: [Farms]
      operationId: replaceFarm
      summary: Replace farm
      x-permission: farm.profile.manage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FarmInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Farm' }
    patch:
      tags: [Farms]
      operationId: updateFarm
      summary: Update farm
      x-permission: farm.profile.manage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FarmInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Farm' }
        '422': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Farms]
      operationId: closeFarm
      x-permission: farm.delete
      summary: Close (archive) the farm. Owner only; data is retained.
      responses:
        '204': { description: 'Closed' }
        '403': { $ref: '#/components/responses/Problem' }
...
  /farms/{farm}/seasons/{season}:
    parameters:
      - $ref: '#/components/parameters/Farm'
      - { name: season, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [Crops]
      operationId: updateSeason
      summary: Edit a season
      x-permission: crops.plans.manage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SeasonInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Season' }
        '422': { $ref: '#/components/responses/Problem' }
...
  /farms/{farm}/crop-cycles/{cycle}:
    parameters:
      - $ref: '#/components/parameters/Farm'
      - { name: cycle, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Crops]
      operationId: getCropCycle
      summary: Get a crop cycle
      x-permission: crops.plans.view
      responses:
        '200':
          description: The cycle
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/CropCycle' }
    patch:
      tags: [Crops]
      operationId: updateCropCycle
      summary: Edit expected dates, yield, area or notes
      x-permission: crops.plans.manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                expected_harvest_on: { type: [string, 'null'], format: date }
                expected_yield: { type: [number, 'null'] }
                yield_unit: { type: string }
                area_ha: { type: number }
                seedlings_germinated: { type: [integer, 'null'] }
                notes: { type: [string, 'null'] }
                version: { type: integer }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/CropCycle' }
        '409': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Farm:
      name: farm
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: 'RFC 9457 problem details'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: Stable machine-readable code (e.g. validation_failed, not_found, mfa_setup_required, token_reuse_detected, append_only)
        errors:
          type: object
          additionalProperties: { type: array, items: { type: string } }
        request_id: { type: [string, 'null'] }
...
    FarmInput:
      type: object
      properties:
        name: { type: string, maxLength: 150 }
        organization_name: { type: string, maxLength: 150 }
        district: { type: [string, 'null'] }
        village: { type: [string, 'null'] }
        country: { type: string, minLength: 2, maxLength: 2 }
        size_ha: { type: [number, 'null'], minimum: 0 }
        timezone: { type: string }
        currency: { type: string, minLength: 3, maxLength: 3 }
    Farm:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, const: farm }
        organization_id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        status: { type: string, enum: [pending, active, suspended, closed] }
        district: { type: [string, 'null'] }
        village: { type: [string, 'null'] }
        country: { type: string }
        size_ha: { type: [string, 'null'], description: 'decimal string' }
        timezone: { type: string }
        currency: { type: string }
        approved_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    CropCycleStage:
      type: string
      enum: [nursery, planted, growing, harvesting, closed]
...
    Ref:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
    BatchRef:
      oneOf:
        - type: 'null'
        - type: object
          properties:
            id: { type: string, format: uuid }
            batch_code: { type: string }
            status: { type: string }
...
    Season:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, const: crop_season }
        name: { type: string }
        starts_on: { type: string, format: date }
        ends_on: { type: string, format: date }
        notes: { type: [string, 'null'] }
    SeasonInput:
      type: object
      properties:
        name: { type: string, maxLength: 80 }
        starts_on: { type: string, format: date }
        ends_on: { type: string, format: date }
        notes: { type: [string, 'null'] }
...
    CropCycle:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, const: crop_cycle }
        code: { type: string }
        stage: { $ref: '#/components/schemas/CropCycleStage' }
        planting_method: { type: string, enum: [direct, transplant] }
        plot: { $ref: '#/components/schemas/Ref' }
        crop:
          type: object
          properties:
            id: { type: string, format: uuid }
            label: { type: string }
        plan: { oneOf: [{ $ref: '#/components/schemas/Ref' }, { type: 'null' }] }
        season:
          oneOf:
            - type: 'null'
            - type: object
              properties:
                id: { type: string, format: uuid }
                name: { type: string }
        area_ha: { type: number }
        sown_on: { type: [string, 'null'], format: date }
        planted_on: { type: [string, 'null'], format: date }
        expected_harvest_on: { type: [string, 'null'], format: date }
        safe_harvest_on: { type: [string, 'null'], format: date, description: 'End of the latest withholding period' }
        expected_yield: { type: [number, 'null'] }
        actual_yield: { type: number, description: 'Harvested so far, in yield_unit' }
        yield_unit: { type: string }
        nursery:
          oneOf:
            - type: 'null'
            - type: object
              properties:
                seeds_sown: { type: [integer, 'null'] }
                seedlings_germinated: { type: [integer, 'null'] }
                seedlings_transplanted: { type: [integer, 'null'] }
        seed_batch: { $ref: '#/components/schemas/BatchRef' }
        nursery_batch: { $ref: '#/components/schemas/BatchRef' }
        crop_lot: { $ref: '#/components/schemas/BatchRef' }
        open_observations: { type: integer }
        operations_count: { type: integer }
        closed_on: { type: [string, 'null'], format: date }
        close_reason: { type: [string, 'null'], enum: [harvested, failed, abandoned, other, null] }
        notes: { type: [string, 'null'] }
        version: { type: integer }
        created_at: { type: string, format: date-time }
...